Refactor code into gui package
